Added support for Windows and the font-awesome icons that are used in the browser and Quazip by modifying the compile settings for both, and by using the actual icon inside the program instead of their unicode equivalents.  Also made some changes to remove some warnings about size_t to int conversions.

+++ src/lib/work_items/bulk_work_item.h
@@ -151,7 +151,7 @@ inline size_t
 BulkWorkItem::GetNumChunksProcessed() const
 {
 	m_responseLock.lock();
-	int chunks = m_numChunksProcessed;
+	size_t chunks = m_numChunksProcessed;
 	m_responseLock.unlock();
 	return chunks;
 }
